Unity Project for LiteNetLib-PacketExample
This is a unity project that uses a client and server, connects them on same machine, then using space bar the client sends a message to the server and pressing enter will have the server send a message to the client. There is an error with the NetPacketProcessor when each of them receive the packet.